refactor(substrate-bundle): split desktop driver event and render concerns
