+++ crates/aether-substrate-bundle/src/desktop/driver.rs
@@ -18,7 +18,7 @@
 
 use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
-use std::time::Instant;
+use std::time::{Duration, Instant};
 
 use aether_actor::Addressable;
 use aether_capabilities::InputCapability;
@@ -26,26 +26,18 @@ use aether_capabilities::RenderHandles;
 use aether_data::Kind;
 use aether_data::{encode, encode_empty, mailbox_id_from_name};
 use aether_kinds::{
-    CaptureFrameResult, FocusWindow, FocusWindowResult, ImePreedit, Key, KeyRelease,
-    LifecycleAdvanceComplete, Modifiers, MouseButton, MouseButtonRelease, MouseMove, MouseWheel,
-    Quit, SetWindowMode, SetWindowModeResult, SetWindowTitle, SetWindowTitleResult, TextInput,
-    Tick, WindowMode, WindowSize, keycode, mouse_button,
+    CaptureFrameResult, FocusWindow, FocusWindowResult, ImePreedit, Key, KeyRelease, Modifiers,
+    MouseButton, MouseButtonRelease, MouseMove, MouseWheel, Quit, SetWindowMode,
+    SetWindowModeResult, SetWindowTitle, SetWindowTitleResult, TextInput, Tick, WindowMode,
+    WindowSize,
 };
 use aether_substrate::actor::native::local;
-// Only the unit tests name the `Envelope` (aka `OwnedDispatch`) type now —
-// `try_framework_dispatch` reaches it through `InboundMail::envelope()`.
-#[cfg(test)]
-use aether_substrate::actor::native::envelope::Envelope;
-use aether_substrate::actor::native::{
-    dispatch_cost_tail_if_matching_free, dispatch_log_tail_if_matching_free,
-    dispatch_trace_tail_if_matching_free,
-};
 use aether_substrate::chassis::builder::{DriverCapability, DriverCtx, DriverRunning, RunError};
 use aether_substrate::chassis::error::BootError;
 use aether_substrate::chassis::settlement::{
     TerminalDisposition, WaitOutcome, await_internal_signal,
 };
-use aether_substrate::runtime::lifecycle;
+use aether_substrate::runtime::lifecycle as runtime_lifecycle;
 use aether_substrate::{
     HubOutbound, InboundMail, Mailer, SettlingInbox, SharedActorSlots, Source, SourceAddr,
     SubstrateBoot,
@@ -53,20 +45,29 @@ use aether_substrate::{
     mail::{Mail, MailId, MailboxId},
 };
 use winit::application::ApplicationHandler;
-use winit::event::{
-    ElementState, Ime, MouseButton as WinitMouseButton, MouseScrollDelta, WindowEvent,
-};
-use winit::event_loop::{ActiveEventLoop, ControlFlow, EventLoop, EventLoopProxy};
-use winit::keyboard::{KeyCode, PhysicalKey};
-use winit::monitor::{MonitorHandle, VideoModeHandle};
-use winit::window::{Fullscreen, Window, WindowId};
+use winit::event::{ElementState, Ime, WindowEvent};
+use winit::event_loop::{ActiveEventLoop, ControlFlow, EventLoop};
+use winit::keyboard::PhysicalKey;
+use winit::window::{Window, WindowId};
 
 use super::chassis::UserEvent;
 use super::render::Gpu;
-use aether_substrate::capture::{CaptureQueue, PendingCapture};
+use aether_substrate::capture::CaptureQueue;
+use capture::{OccludedCaptureDisposition, occluded_capture_disposition};
+use config::resolve_fullscreen;
+use input::{TextSource, map_mouse_button, map_winit_keycode, normalize_wheel, text_input_gate};
+use lifecycle::{LifecycleReplyOutcome, consume_lifecycle_reply, try_framework_dispatch};
+use shutdown::install_shutdown_handler;
 use std::io;
-use std::time::Duration;
 use winit::dpi::PhysicalSize;
+
+mod capture;
+mod config;
+mod input;
+mod lifecycle;
+mod shutdown;
+
+pub use config::{WindowConfig, WindowConfigLayer, WindowOverlay, parse_window_mode_env};
 
 /// Cumulative patience cap for the per-frame settlement gates (advance +
 /// capture pre-mail). The per-round budget is `frame_loop::DRAIN_BUDGET`
@@ -232,547 +233,6 @@ pub struct App {
     /// struct field (mirroring headless's flag) so the watcher and the
     /// winit loop share one source of truth.
     shutdown: Arc<AtomicBool>,
-}
-
-/// Translate a winit `KeyCode` into the engine's stable named-key u32
-/// space (`aether_kinds::keycode`). Returns `None` for any key the
-/// engine doesn't name yet — the event then drops at the source rather
-/// than leaking winit's unstable discriminants onto the wire. Adding
-/// a new key is a paired change: a constant in `aether-kinds::keycode`
-/// plus an arm here.
-fn map_winit_keycode(k: KeyCode) -> Option<u32> {
-    Some(match k {
-        KeyCode::KeyA => keycode::KEY_A,
-        KeyCode::KeyB => keycode::KEY_B,
-        KeyCode::KeyC => keycode::KEY_C,
-        KeyCode::KeyD => keycode::KEY_D,
-        KeyCode::KeyE => keycode::KEY_E,
-        KeyCode::KeyF => keycode::KEY_F,
-        KeyCode::KeyG => keycode::KEY_G,
-        KeyCode::KeyH => keycode::KEY_H,
-        KeyCode::KeyI => keycode::KEY_I,
-        KeyCode::KeyJ => keycode::KEY_J,
-        KeyCode::KeyK => keycode::KEY_K,
-        KeyCode::KeyL => keycode::KEY_L,
-        KeyCode::KeyM => keycode::KEY_M,
-        KeyCode::KeyN => keycode::KEY_N,
-        KeyCode::KeyO => keycode::KEY_O,
-        KeyCode::KeyP => keycode::KEY_P,
-        KeyCode::KeyQ => keycode::KEY_Q,
-        KeyCode::KeyR => keycode::KEY_R,
-        KeyCode::KeyS => keycode::KEY_S,
-        KeyCode::KeyT => keycode::KEY_T,
-        KeyCode::KeyU => keycode::KEY_U,
-        KeyCode::KeyV => keycode::KEY_V,
-        KeyCode::KeyW => keycode::KEY_W,
-        KeyCode::KeyX => keycode::KEY_X,
-        KeyCode::KeyY => keycode::KEY_Y,
-        KeyCode::KeyZ => keycode::KEY_Z,
-        KeyCode::Digit0 => keycode::KEY_0,
-        KeyCode::Digit1 => keycode::KEY_1,
-        KeyCode::Digit2 => keycode::KEY_2,
-        KeyCode::Digit3 => keycode::KEY_3,
-        KeyCode::Digit4 => keycode::KEY_4,
-        KeyCode::Digit5 => keycode::KEY_5,
-        KeyCode::Digit6 => keycode::KEY_6,
-        KeyCode::Digit7 => keycode::KEY_7,
-        KeyCode::Digit8 => keycode::KEY_8,
-        KeyCode::Digit9 => keycode::KEY_9,
-        KeyCode::Backquote => keycode::KEY_BACKQUOTE,
-        KeyCode::Space => keycode::KEY_SPACE,
-        KeyCode::Escape => keycode::KEY_ESCAPE,
-        KeyCode::Enter => keycode::KEY_ENTER,
-        KeyCode::Tab => keycode::KEY_TAB,
-        KeyCode::Backspace => keycode::KEY_BACKSPACE,
-        KeyCode::ArrowLeft => keycode::KEY_LEFT,
-        KeyCode::ArrowRight => keycode::KEY_RIGHT,
-        KeyCode::ArrowUp => keycode::KEY_UP,
-        KeyCode::ArrowDown => keycode::KEY_DOWN,
-        KeyCode::ShiftLeft => keycode::KEY_SHIFT_LEFT,
-        KeyCode::ShiftRight => keycode::KEY_SHIFT_RIGHT,
-        KeyCode::ControlLeft => keycode::KEY_CTRL_LEFT,
-        KeyCode::ControlRight => keycode::KEY_CTRL_RIGHT,
-        KeyCode::AltLeft => keycode::KEY_ALT_LEFT,
-        KeyCode::AltRight => keycode::KEY_ALT_RIGHT,
-        _ => return None,
-    })
-}
-
-/// A committed-text / composition signal lifted out of winit's event
-/// types so [`text_input_gate`]'s dedupe logic is unit-testable without
-/// a winit event loop — the same pure-helper factoring as
-/// [`map_winit_keycode`]. The desktop `KeyboardInput` / `Ime` arms
-/// translate their winit events into this before feeding the gate.
-enum TextSource {
-    /// A layout-resolved character run from `KeyEvent.text` on a physical
-    /// key press. Suppressed while an IME composition is active.
-    KeyText(String),
-    /// A `Ime::Preedit`. `active` is `true` for a non-empty preedit
-    /// (composition open) and `false` for the synthetic empty preedit
-    /// winit sends to clear it.
-    Preedit { active: bool },
-    /// A `Ime::Commit` — the composed string is final.
-    Commit(String),
-    /// A `Ime::Disabled` — composition ended without a commit.
-    Disabled,
-}
-
-/// Composition gate for the `TextInput` stream: update `composing` and
-/// return the text to publish (`Some`) or nothing (`None`). The bug it
-/// guards is a character delivered twice — once as `KeyEvent.text`, once
-/// as `Ime::Commit` — when both fire for one keystroke: while a
-/// composition is open, `KeyText` is dropped and the commit is the single
-/// source of truth. `Preedit { active }` opens or closes the gate without
-/// publishing (the in-flight text rides `ImePreedit` instead); `Disabled`
-/// closes it. Winit-free so the dedupe is testable without a winit `App`.
-///
-/// `KeyText` also strips control characters before publishing. Winit
-/// reports a named key's (Backspace / Enter / Tab / Escape / Delete) text
-/// representation as a C0 control character, and those keys already
-/// arrive as `Key` scancode edges — publishing the control character too
-/// would double-report the keystroke as printable text (e.g. Backspace
-/// inserting a glyph on the same frame its scancode edge deletes one).
-/// The strip is per-character rather than whole-event: winit documents
-/// that `KeyEvent.text` can carry a dead-key char followed by a resolved
-/// character, so a run is not guaranteed to be a single glyph, and
-/// dropping only the control chars preserves any printable characters
-/// riding alongside them.
-fn text_input_gate(composing: &mut bool, source: TextSource) -> Option<String> {
-    match source {
-        TextSource::KeyText(text) => (!*composing)
-            .then(|| text.chars().filter(|c| !c.is_control()).collect::<String>())
-            .filter(|t| !t.is_empty()),
-        TextSource::Preedit { active } => {
-            *composing = active;
-            None
-        }
-        TextSource::Commit(text) => {
-            *composing = false;
-            Some(text)
-        }
-        TextSource::Disabled => {
-            *composing = false;
-            None
-        }
-    }
-}
-
-/// Desktop window boot knobs (ADR-0090 §1/§2 applied to the chassis's
-/// own window knobs). The `#[derive(aether_substrate::Config)]` emits
-/// the env-shaped `WindowConfigLayer`, the clap-shaped `WindowOverlay`,
-/// the `FromArgvThenEnv` impl, and the inherent `from_env` /
-/// `from_argv_then_env` shims — mirrors `ActorRingConfig` in `chassis_common`.
-///
-/// `env_prefix = "AETHER_WINDOW"` joins the field env keys — `mode` →
-/// `AETHER_WINDOW_MODE`, `title` → `AETHER_WINDOW_TITLE` — matching the
-/// historical key names exactly without per-field overrides. Both fields
-/// are `Option<String>` (soft: missing or empty → `None`).
-#[derive(Clone, Debug, Default, aether_substrate::Config)]
-#[config(env_prefix = "AETHER_WINDOW", cli_prefix = "window")]
-pub struct WindowConfig {
-    /// `AETHER_WINDOW_MODE=<value>` desktop window mode at boot:
-    /// `windowed[:WxH]` / `fullscreen-borderless` / `exclusive:WxH@HZ`.
-    /// Lowered via [`Self::to_boot_mode`] which delegates to
-    /// [`parse_window_mode_env`] and soft-falls back to `Windowed` on
-    /// a bad value (keeping the pre-migration behaviour).
-    pub mode: Option<String>,
-    /// `AETHER_WINDOW_TITLE=<text>` desktop window title at boot.
-    /// Lowered via [`Self::to_boot_title`]; empty / unset → `"aether"`.
-    pub title: Option<String>,
-    /// `AETHER_WIREFRAME=<value>` desktop GPU wireframe mode at boot:
-    /// `off` (default) / `line` / `overlay`. The env key is pinned back
-    /// to `AETHER_WIREFRAME` (not `AETHER_WINDOW_WIREFRAME`) and the
-    /// argv flag to `--wireframe` (not `--window-wireframe`) since the
-    /// knob predates joining `WindowConfig`. Threaded verbatim to
-    /// `render::WireframeMode::from_config_value`, which owns the
-    /// tri-state parse.
-    #[config(env = "AETHER_WIREFRAME", cli_long = "wireframe")]
-    pub wireframe: Option<String>,
-}
-
-impl WindowConfig {
-    /// Lower `mode` to the `(WindowMode, Option<(u32, u32)>)` pair the
-    /// chassis threads into `DesktopEnv`. Delegates to
-    /// [`parse_window_mode_env`]; a bad value warn-logs and falls back
-    /// to `Windowed`, preserving the pre-migration soft-fallback
-    /// behaviour for `AETHER_WINDOW_MODE`.
-    #[must_use]
-    pub fn to_boot_mode(&self) -> (WindowMode, Option<(u32, u32)>) {
-        self.mode.as_ref().map_or(
-            (WindowMode::Windowed, None),
-            |s| match parse_window_mode_env(s) {
-                Ok(parsed) => parsed,
-                Err(e) => {
-                    tracing::warn!(
-                        target: "aether_substrate::boot",
-                        value = %s,
-                        error = %e,
-                        "AETHER_WINDOW_MODE unparseable — falling back to Windowed",
-                    );
-                    (WindowMode::Windowed, None)
-                }
-            },
-        )
-    }
-
-    /// Lower `title` to the boot title string. `None` (unset or empty —
-    /// the derive's `Option<String>` filters empty → `None`) maps to
-    /// `"aether"`; a provided value passes through verbatim.
-    #[must_use]
-    pub fn to_boot_title(&self) -> String {
-        self.title.clone().unwrap_or_else(|| "aether".to_owned())
-    }
-}
-
-/// Parse `AETHER_WINDOW_MODE`. Grammar:
-///   `windowed`              — default size
-///   `windowed:WxH`          — windowed, `WxH` physical pixels
-///   `fullscreen-borderless` — borderless on current monitor
-///   `exclusive:WxH@HZ`      — exclusive, matched against monitor modes
-/// Refresh is integer Hz (converted to mhz by *1000); non-integer
-/// refresh isn't expressible from the env var today — runtime
-/// `set_window_mode` accepts full-precision mhz directly.
-pub fn parse_window_mode_env(s: &str) -> Result<(WindowMode, Option<(u32, u32)>), String> {
-    let s = s.trim();
-    if s == "windowed" {
-        return Ok((WindowMode::Windowed, None));
-    }
-    if let Some(rest) = s.strip_prefix("windowed:") {
-        let (w, h) = parse_wxh(rest)?;
-        return Ok((WindowMode::Windowed, Some((w, h))));
-    }
-    if s == "fullscreen-borderless" {
-        return Ok((WindowMode::FullscreenBorderless, None));
-    }
-    if let Some(rest) = s.strip_prefix("exclusive:") {
-        let (dim, hz) = rest
-            .split_once('@')
-            .ok_or_else(|| format!("exclusive mode missing @HZ in {s:?}"))?;
-        let (width, height) = parse_wxh(dim)?;
-        let hz: u32 = hz.parse().map_err(|e| format!("invalid Hz {hz:?}: {e}"))?;
-        return Ok((
-            WindowMode::FullscreenExclusive {
-                width,
-                height,
-                refresh_mhz: hz.saturating_mul(1000),
-            },
-            None,
-        ));
-    }
-    Err(format!("unrecognised AETHER_WINDOW_MODE value {s:?}"))
-}
-
-fn parse_wxh(s: &str) -> Result<(u32, u32), String> {
-    let (w, h) = s
-        .split_once('x')
-        .ok_or_else(|| format!("expected WxH, got {s:?}"))?;
-    let w: u32 = w.parse().map_err(|e| format!("invalid width {w:?}: {e}"))?;
-    let h: u32 = h
-        .parse()
-        .map_err(|e| format!("invalid height {h:?}: {e}"))?;
-    Ok((w, h))
-}
-
-/// Scroll lines are normalized to pixels at this rate. A tuning knob,
-/// not cap config — the wheel kind carries pixel-space deltas so a
-/// consumer never sees the winit line/pixel distinction.
-const PIXELS_PER_SCROLL_LINE: f32 = 40.0;
-
-/// Map winit's mouse button to the engine's `mouse_button` constant
-/// space. `Other(n)` maps to `None` — the caller pushes no mail,
-/// mirroring the unmapped-key contract in `keycode`.
-fn map_mouse_button(button: WinitMouseButton) -> Option<u32> {
-    match button {
-        WinitMouseButton::Left => Some(mouse_button::LEFT),
-        WinitMouseButton::Right => Some(mouse_button::RIGHT),
-        WinitMouseButton::Middle => Some(mouse_button::MIDDLE),
-        WinitMouseButton::Back => Some(mouse_button::BACK),
-        WinitMouseButton::Forward => Some(mouse_button::FORWARD),
-        WinitMouseButton::Other(_) => None,
-    }
-}
-
-/// Normalize a winit scroll delta to pixel-space `(delta_x, delta_y)`.
-/// Line deltas scale by `PIXELS_PER_SCROLL_LINE`; pixel deltas cast
-/// `f64 → f32` directly, matching the `CursorMoved` position cast.
-fn normalize_wheel(delta: MouseScrollDelta) -> (f32, f32) {
-    match delta {
-        MouseScrollDelta::LineDelta(x, y) => {
-            (x * PIXELS_PER_SCROLL_LINE, y * PIXELS_PER_SCROLL_LINE)
-        }
-        // Realistic scroll deltas stay well inside f32 mantissa.
-        #[allow(clippy::cast_possible_truncation)]
-        MouseScrollDelta::PixelDelta(p) => (p.x as f32, p.y as f32),
-    }
-}
-
-/// Find a `VideoModeHandle` on `monitor` matching the given size +
-/// refresh exactly. Returns `None` if no match — the caller surfaces
-/// this as `SetWindowModeResult::Err` rather than falling back
-/// silently to something close.
-fn find_exclusive_mode(
-    monitor: &MonitorHandle,
-    width: u32,
-    height: u32,
-    refresh_mhz: u32,
-) -> Option<VideoModeHandle> {
-    monitor.video_modes().find(|m| {
-        m.size().width == width
-            && m.size().height == height
-            && m.refresh_rate_millihertz() == refresh_mhz
-    })
-}
-
-/// Build winit's `Option<Fullscreen>` for the requested mode.
-/// `monitor_for_exclusive` is the monitor to match video modes
-/// against — the window's current monitor at runtime, the primary at
-/// boot.
-fn resolve_fullscreen(
-    mode: &WindowMode,
-    monitor_for_exclusive: Option<&MonitorHandle>,
-) -> Result<Option<Fullscreen>, String> {
-    match mode {
-        WindowMode::Windowed => Ok(None),
-        WindowMode::FullscreenBorderless => Ok(Some(Fullscreen::Borderless(None))),
-        WindowMode::FullscreenExclusive {
-            width,
-            height,
-            refresh_mhz,
-        } => {
-            let monitor = monitor_for_exclusive.ok_or_else(|| {
-                "fullscreen-exclusive requested but no monitor available".to_owned()
-            })?;
-            let handle =
-                find_exclusive_mode(monitor, *width, *height, *refresh_mhz).ok_or_else(|| {
-                    format!(
-                        "no video mode matches {width}x{height}@{refresh_mhz}mhz on monitor {:?}",
-                        monitor.name()
-                    )
-                })?;
-            Ok(Some(Fullscreen::Exclusive(handle)))
-        }
-    }
-}
-
-/// iamacoffeepot/aether#1272 / #1710: route an inbound `aether.window`
-/// envelope through the framework-built-in dispatch arms
-/// (`aether.log.tail` / `aether.trace.tail` / `aether.cost.tail`) before
-/// the driver-specific `SetWindowMode` / `SetWindowTitle` arms get their
-/// turn. Each helper computes-and-returns its `*TailResult`; on a match
-/// the reply rides the inbound's drain guard — [`InboundMail::reply`]
-/// mints the reply id on the drain-owned counter and stamps the
-/// inbound's `root` / parent, so the framework-arm reply joins the
-/// caller's ADR-0080 chain exactly like the window-mode / title arms.
-/// Returns `true` when one of the framework arms matched (the reply has
-/// been routed); `false` otherwise. ADR-0081 §1 promises every mailbox
-/// serves these kinds — see the issue body for the contract.
-///
-/// Caller invariant: must run inside a `local::with_stamped` block
-/// against the driver's [`local::ActorSlots`] so the
-/// log / trace arms reach the driver's per-actor ring. Factored out of
-/// [`App::dispatch_window_envelope`] so the unit test directly drives
-/// the routing shape without standing up a winit `App`.
-fn try_framework_dispatch(
-    mailer: &Arc<Mailer>,
-    self_mailbox: MailboxId,
-    mail: &InboundMail,
-) -> bool {
-    let env = mail.envelope();
-    if let Some(result) = dispatch_log_tail_if_matching_free(env) {
-        mail.reply(&result);
-        return true;
-    }
-    if let Some(result) = dispatch_trace_tail_if_matching_free(env) {
-        mail.reply(&result);
-        return true;
-    }
-    if let Some(result) = dispatch_cost_tail_if_matching_free(mailer.as_ref(), self_mailbox, env) {
-        mail.reply(&result);
-        return true;
-    }
-    false
-}
-
-/// The disposition of one consumed `aether.lifecycle.advance_reply`
-/// envelope, returned by [`consume_lifecycle_reply`] to
-/// [`App::recv_lifecycle_advance_next`].
-enum LifecycleReplyOutcome {
-    /// The envelope was the expected [`LifecycleAdvanceComplete`]. Carries
-    /// the decoded `next` stage kind id, or `None` when the payload failed
-    /// to decode — the caller fail-fasts that the same as a missing reply.
-    Complete(Option<u64>),
-    /// An unexpected kind on the dedicated reply inbox (nothing else
-    /// targets it). Discharged like the matched arm; the caller keeps
-    /// waiting for the advance reply rather than mis-gating the cycle.
-    Unexpected,
-}
-
-/// Consume one [`InboundMail`] off the lifecycle reply inbox. The mail's
-/// ADR-0094 obligation guard + ADR-0080 §2 settlement bracket discharge
-/// when the guard falls out of scope (ADR-0106) — the same scope-exit
-/// settle [`App::dispatch_window_envelope`] relies on for the sibling
-/// `aether.window` inbox. The hand-rolled per-arm `record_finished` +
-/// `discharge()` pairs that #1325 / #1704 added retired with the
-/// framework drain: dropping `mail` on either arm settles.
-///
-/// On the per-frame path the reply rides a bare, lineage-less `Settled`
-/// notice, so its `root` is `MailId::NONE` and the drop's `record_finished`
-/// is a counter no-op; the live obligation it discharges is the debug
-/// guard the real `route_mail` Inbox arm armed.
-//
-// `mail` is taken by value so its guard's `Drop` (the settlement) binds
-// to this scope; the body only calls `&self` accessors, which clippy
-// reads as a needless by-value.
-#[allow(clippy::needless_pass_by_value)]
-fn consume_lifecycle_reply(mail: InboundMail) -> LifecycleReplyOutcome {
-    if mail.kind() == <LifecycleAdvanceComplete as Kind>::ID {
-        LifecycleReplyOutcome::Complete(
-            LifecycleAdvanceComplete::decode_from_bytes(mail.payload())
-                .map(|complete| complete.next),
-        )
-    } else {
-        LifecycleReplyOutcome::Unexpected
-    }
-    // `mail` drops here — both arms settle (ADR-0106).
-}
-
-/// The disposition of a `capture_frame` wake against the current window
-/// occlusion state — the pure branch-selection half of the occluded
-/// fail-fast (iamacoffeepot/aether#1317), factored out of the winit
-/// wiring so it is unit-testable without standing up a winit `App`
-/// (mirroring [`try_framework_dispatch`]).
-enum OccludedCaptureDisposition {
-    /// Window is visible — the caller falls through to `request_redraw`
-    /// so `RedrawRequested` services the (still-parked) capture normally.
-    Redraw,
-    /// Window is occluded but no capture was parked (already serviced, or
-    /// a stale wake) — nothing to do.
-    Empty,
-    /// Window is occluded and a capture is parked — fail it fast. Carries
-    /// the taken [`PendingCapture`] (so the caller drains `after_mails`,
-    /// replies through its retained inbound guard, then drops it to settle
-    /// the inbound chain) and the `Err` reply to send.
-    FailFast {
-        // Boxed: `PendingCapture` carries a retained `InboundMail` guard, so
-        // it dwarfs the zero-size `Redraw` / `Empty` variants
-        // (clippy::large_enum_variant).
-        request: Box<PendingCapture>,
-        result: CaptureFrameResult,
-    },
-}
-
-/// The capture-frame error message naming `aether.window.focus` as the
-/// remedy. Shared by every occluded fail-fast site so the wording can't
-/// drift.
-const OCCLUDED_CAPTURE_ERROR: &str = "capture_frame: window is occluded (hidden/minimized); bring it to the \
-     foreground via aether.window.focus and retry";
-
-/// Select the disposition for a `capture_frame` wake (or an occlusion
-/// onset) given the window's occlusion state and any parked capture.
-///
-/// The winit side only `take()`s the [`CaptureQueue`] slot when occluded,
-/// so a visible-window wake never steals the entry that `RedrawRequested`
-/// is about to service — the `Redraw` arm carries no `PendingCapture`. The
-/// occluded arms move the taken request through so the caller can drain
-/// `after_mails`, reply through the retained inbound guard, and drop the
-/// request *after* the reply (settling the inbound chain post-reply,
-/// ADR-0080 §6 / iamacoffeepot/aether#1273).
-fn occluded_capture_disposition(
-    occluded: bool,
-    pending: Option<PendingCapture>,
-) -> OccludedCaptureDisposition {
-    if !occluded {
-        return OccludedCaptureDisposition::Redraw;
-    }
-    pending.map_or(OccludedCaptureDisposition::Empty, |request| {
-        OccludedCaptureDisposition::FailFast {
-            request: Box::new(request),
-            result: CaptureFrameResult::Err {
-                error: OCCLUDED_CAPTURE_ERROR.to_owned(),
-            },
-        }
-    })
-}
-
-/// Install a SIGINT/SIGTERM → graceful-shutdown bridge for the desktop
-/// chassis (iamacoffeepot/aether#1489). On the first delivered signal it
-/// flips `shutdown` (the flag [`App::about_to_wait`] polls) and sends
-/// [`UserEvent::Quit`] through `proxy` to wake a parked
-/// (`ControlFlow::Wait`, occluded) loop — the loop then runs
-/// `about_to_wait`, observes the flag, and drives the lifecycle to
-/// `Shutdown` (the desktop analogue of headless's tick-loop flag poll).
-///
-/// Unlike headless's `signal_hook::flag::register` — which is
-/// async-signal-safe but can only flip a bool — the desktop loop must be
-/// *woken*, and `EventLoopProxy::send_event` is not async-signal-safe.
-/// So a dedicated watcher thread blocks on the signal stream and does
-/// both the flag flip and the proxy wake; it doesn't freeze the winit
-/// loop (a separate thread), so the constraint that ruled this out for
-/// the single-threaded headless tick loop doesn't apply here. SIGTERM
-/// joins SIGINT so supervisors / `kill` (no `-9`) / CI cancellation also
-/// run teardown. Best-effort: a failed install warn-logs and leaves
-/// shutdown to `WindowEvent::CloseRequested` only.
-#[cfg(unix)]
-fn install_shutdown_handler(shutdown: &Arc<AtomicBool>, proxy: EventLoopProxy<UserEvent>) {
-    use std::thread;
-
-    use signal_hook::consts::{SIGINT, SIGTERM};
-    use signal_hook::iterator::Signals;
-
-    let shutdown = Arc::clone(shutdown);
-    let mut signals = match Signals::new([SIGINT, SIGTERM]) {
-        Ok(s) => s,
-        Err(e) => {
-            tracing::error!(
-                target: "aether_substrate::boot",
-                error = %e,
-                "desktop: shutdown signal handler install failed; \
-                 only window-close will trigger graceful shutdown",
-            );
-            return;
-        }
-    };
-    // Infra thread: it blocks on the OS signal stream and holds no
-    // settlement/trace contract — the work it triggers (the `Quit` push)
-    // happens later, on the winit main thread, through the normal mail
-    // path. A separate thread (not the single-threaded winit loop), so
-    // it never freezes the loop.
-    #[allow(clippy::disallowed_methods)]
-    let spawned = thread::Builder::new()
-        .name("aether-desktop-signal".into())
-        .spawn(move || {
-            // The first signal begins graceful shutdown; the iterator
-            // only ends if the underlying fd closes (it doesn't for the
-            // thread's lifetime), so a single `next()` is the whole job.
-            if signals.forever().next().is_some() {
-                shutdown.store(true, Ordering::SeqCst);
-                let _ = proxy.send_event(UserEvent::Quit);
-            }
-        });
-    if let Err(e) = spawned {
-        tracing::error!(
-            target: "aether_substrate::boot",
-            error = %e,
-            "desktop: shutdown signal-watcher thread failed to spawn; \
-             only window-close will trigger graceful shutdown",
-        );
-    }
-}
-
-#[cfg(not(unix))]
-fn install_shutdown_handler(shutdown: &Arc<AtomicBool>, proxy: EventLoopProxy<UserEvent>) {
-    let shutdown = Arc::clone(shutdown);
-    if let Err(e) = ctrlc::set_handler(move || {
-        shutdown.store(true, Ordering::SeqCst);
-        let _ = proxy.send_event(UserEvent::Quit);
-    }) {
-        tracing::error!(
-            target: "aether_substrate::boot",
-            error = %e,
-            "desktop: ctrl-c handler install failed; \
-             only window-close will trigger graceful shutdown",
-        );
-    }
 }
 
 impl App {
@@ -1283,7 +743,7 @@ impl ApplicationHandler<UserEvent> for App {
                             FRAME_SETTLEMENT_CAP,
                             TerminalDisposition::Abort,
                         ) {
-                            lifecycle::fatal_abort(&self.outbound, wedge.reason());
+                            runtime_lifecycle::fatal_abort(&self.outbound, wedge.reason());
                         }
                     }
                     match self.recv_lifecycle_advance_next() {
@@ -1302,7 +762,7 @@ impl ApplicationHandler<UserEvent> for App {
                         // Settlement fired but the reply never arrived —
                         // a wedge in the reply path; fail-fast like the
                         // settlement wait above.
-                        None => lifecycle::fatal_abort(
+                        None => runtime_lifecycle::fatal_abort(
                             &self.outbound,
                             "desktop.frame_advance: LifecycleAdvanceComplete reply did not \
                              arrive after settlement"
@@ -1804,834 +1264,5 @@ impl DriverRunning for DesktopDriverRunning {
             "frame loop exited",
         );
         Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    // Tests derive chassis mailbox ids by name to address window/lifecycle mail
-    // in fixtures — reference id derivation, not sibling-cap addressing.
-    #![allow(clippy::disallowed_methods)]
-    use super::*;
-    use winit::dpi::PhysicalPosition;
-
-    #[test]
-    fn to_boot_title_none_returns_default() {
-        // Unset title → "aether" default.
-        assert_eq!(WindowConfig::default().to_boot_title(), "aether");
-    }
-
-    // Tripwire: the composition gate must never publish a character twice.
-    // A plain keystroke with no IME active publishes its `KeyEvent.text`.
-    #[test]
-    fn gate_publishes_keytext_when_not_composing() {
-        let mut composing = false;
-        let out = text_input_gate(&mut composing, TextSource::KeyText("a".to_owned()));
-        assert_eq!(out.as_deref(), Some("a"));
-        assert!(!composing, "a bare keystroke opens no composition");
-    }
-
-    // Tripwire: while an IME composition is open, raw `KeyEvent.text` is
-    // suppressed — the commit is the single source of truth, so the
-    // committed character is not also emitted from the physical key.
-    #[test]
-    fn gate_suppresses_keytext_during_composition_and_commit_wins() {
-        let mut composing = false;
-        // A non-empty preedit opens the composition.
-        assert_eq!(
-            text_input_gate(&mut composing, TextSource::Preedit { active: true }),
-            None
-        );
-        assert!(composing);
-        // Raw key text arriving mid-composition is dropped.
-        assert_eq!(
-            text_input_gate(&mut composing, TextSource::KeyText("a".to_owned())),
-            None
-        );
-        // The commit publishes the composed text and closes the gate.
-        let out = text_input_gate(&mut composing, TextSource::Commit("\u{5416}".to_owned()));
-        assert_eq!(out.as_deref(), Some("\u{5416}"));
-        assert!(!composing, "commit closes the composition");
-    }
-
-    // Tripwire: text must not stay suppressed after a composition ends.
-    // The synthetic empty preedit and `Disabled` both clear `composing`,
-    // so a subsequent keystroke publishes again.
-    #[test]
-    fn gate_clears_composing_on_empty_preedit_and_disabled() {
-        let mut composing = true;
-        assert_eq!(
-            text_input_gate(&mut composing, TextSource::Preedit { active: false }),
-            None
-        );
-        assert!(!composing, "empty synthetic preedit clears the gate");
-
-        composing = true;
-        assert_eq!(text_input_gate(&mut composing, TextSource::Disabled), None);
-        assert!(!composing, "Disabled clears the gate");
-
-        // A keystroke after the clear publishes normally.
-        assert_eq!(
-            text_input_gate(&mut composing, TextSource::KeyText("z".to_owned())).as_deref(),
-            Some("z"),
-        );
-    }
-
-    // Tripwire: a named key's control-char text representation must never
-    // publish as `TextInput` — Backspace's scancode edge is the sole
-    // delete signal, so a published `"\u{8}"` would re-insert a glyph on
-    // the same frame the edge deletes one.
-    #[test]
-    fn gate_suppresses_pure_backspace_keytext() {
-        let mut composing = false;
-        let out = text_input_gate(&mut composing, TextSource::KeyText("\u{8}".to_owned()));
-        assert_eq!(out, None);
-        assert!(!composing, "a suppressed control char opens no composition");
-    }
-
-    // Tripwire: Enter and Tab carry the same C0-control-char shape as
-    // Backspace and must be suppressed identically.
-    #[test]
-    fn gate_suppresses_pure_enter_and_tab_keytext() {
-        let mut composing = false;
-        assert_eq!(
-            text_input_gate(&mut composing, TextSource::KeyText("\r".to_owned())),
-            None
-        );
-        assert_eq!(
-            text_input_gate(&mut composing, TextSource::KeyText("\t".to_owned())),
-            None
-        );
-    }
-
-    // Tripwire: a run mixing a printable character with a control
-    // character strips only the control char, pinning strip-per-char over
-    // whole-event drop (winit can pair a dead-key char with a resolved
-    // character in one `KeyEvent.text`).
-    #[test]
-    fn gate_strips_control_char_from_mixed_keytext() {
-        let mut composing = false;
-        let out = text_input_gate(&mut composing, TextSource::KeyText("a\u{8}".to_owned()));
-        assert_eq!(out.as_deref(), Some("a"));
-    }
-
-    #[test]
-    fn map_winit_keycode_covers_console_activation_key() {
-        assert_eq!(
-            map_winit_keycode(KeyCode::Backquote),
-            Some(keycode::KEY_BACKQUOTE)
-        );
-    }
-
-    #[test]
-    fn map_mouse_button_covers_named_buttons() {
-        assert_eq!(
-            map_mouse_button(WinitMouseButton::Left),
-            Some(mouse_button::LEFT)
-        );
-        assert_eq!(
-            map_mouse_button(WinitMouseButton::Right),
-            Some(mouse_button::RIGHT)
-        );
-        assert_eq!(
-            map_mouse_button(WinitMouseButton::Middle),
-            Some(mouse_button::MIDDLE)
-        );
-        assert_eq!(
-            map_mouse_button(WinitMouseButton::Back),
-            Some(mouse_button::BACK)
-        );
-        assert_eq!(
-            map_mouse_button(WinitMouseButton::Forward),
-            Some(mouse_button::FORWARD)
-        );
-    }
-
-    #[test]
-    fn map_mouse_button_other_produces_no_mail() {
-        assert_eq!(map_mouse_button(WinitMouseButton::Other(9)), None);
-    }
-
-    #[test]
-    fn map_winit_keycode_covers_backquote() {
-        assert_eq!(
-            map_winit_keycode(KeyCode::Backquote),
-            Some(keycode::KEY_BACKQUOTE)
-        );
-    }
-
-    #[test]
-    fn normalize_wheel_scales_line_delta() {
-        let (x, y) = normalize_wheel(MouseScrollDelta::LineDelta(1.0, -2.0));
-        assert_eq!(x, PIXELS_PER_SCROLL_LINE);
-        assert_eq!(y, -2.0 * PIXELS_PER_SCROLL_LINE);
-    }
-
-    #[test]
-    fn normalize_wheel_passes_pixel_delta_through() {
-        let delta = MouseScrollDelta::PixelDelta(PhysicalPosition::new(12.0, -3.0));
-        let (x, y) = normalize_wheel(delta);
-        assert_eq!(x, 12.0);
-        assert_eq!(y, -3.0);
-    }
-
-    #[test]
-    fn to_boot_title_some_returns_value() {
-        // Provided title passes through verbatim.
-        let cfg = WindowConfig {
-            mode: None,
-            title: Some("my game".to_owned()),
-            wireframe: None,
-        };
-        assert_eq!(cfg.to_boot_title(), "my game");
-    }
-
-    #[test]
-    fn parse_windowed_defaults() {
-        let (m, s) =
-            parse_window_mode_env("windowed").expect("test setup: \"windowed\" is a valid spec");
-        assert!(matches!(m, WindowMode::Windowed));
-        assert_eq!(s, None);
-    }
-
-    #[test]
-    fn parse_windowed_with_size() {
-        let (m, s) = parse_window_mode_env("windowed:1280x720")
-            .expect("test setup: \"windowed:WxH\" is a valid spec");
-        assert!(matches!(m, WindowMode::Windowed));
-        assert_eq!(s, Some((1280, 720)));
-    }
-
-    #[test]
-    fn parse_fullscreen_borderless() {
-        let (m, s) = parse_window_mode_env("fullscreen-borderless")
-            .expect("test setup: \"fullscreen-borderless\" is a valid spec");
-        assert!(matches!(m, WindowMode::FullscreenBorderless));
-        assert_eq!(s, None);
-    }
-
-    #[test]
-    fn parse_exclusive_converts_hz_to_mhz() {
-        let (m, s) = parse_window_mode_env("exclusive:1920x1080@60")
-            .expect("test setup: \"exclusive:WxH@HZ\" is a valid spec");
-        let WindowMode::FullscreenExclusive {
-            width,
-            height,
-            refresh_mhz,
-        } = m
-        else {
-            panic!("expected exclusive");
-        };
-        assert_eq!((width, height, refresh_mhz), (1920, 1080, 60_000));
-        assert_eq!(s, None);
-    }
-
-    #[test]
-    fn parse_rejects_unknown_variant() {
-        assert!(parse_window_mode_env("garbage").is_err());
-        assert!(parse_window_mode_env("exclusive:1920x1080").is_err());
-        assert!(parse_window_mode_env("windowed:notxwide").is_err());
-    }
-
-    #[test]
-    fn parse_ignores_whitespace() {
-        let (m, _) = parse_window_mode_env("  windowed  ")
-            .expect("test setup: surrounding whitespace is trimmed");
-        assert!(matches!(m, WindowMode::Windowed));
-    }
-
-    /// iamacoffeepot/aether#1272 / #1710: a `LogTail` Call drained at the
-    /// driver's `aether.window` mailbox produces a `LogTailResult` reply
-    /// through the inbound's drain guard — and that reply joins the
-    /// inbound's ADR-0080 causal chain (it carries the inbound's `root`)
-    /// rather than minting the lineage-less `MailId::NONE` triple the
-    /// pre-#1710 bare `Mailer::send_reply` form did. Drives a real armed
-    /// Call through a `SettlingInbox` (mirroring
-    /// `window_inbox_drain_settles_root_on_guard_drop`), routes the reply
-    /// at a captured `Component` inbox, and reads the delivered reply's
-    /// `root` — without standing up wgpu/winit.
-    #[test]
-    fn try_framework_dispatch_replies_to_log_tail() {
-        use std::sync::mpsc;
-        use std::time::Duration;
-
-        use aether_actor::local::ActorSlots;
-        use aether_data::MailId;
-        use aether_kinds::descriptors;
-        use aether_kinds::{LogTail, LogTailResult};
-        use aether_substrate::actor::native::local::with_stamped;
-
-        use aether_substrate::chassis::settlement::SettlementRegistry;
-        use aether_substrate::mail::Mail;
-        use aether_substrate::mail::registry::{InboxHandler, OwnedDispatch, Registry};
-        use aether_substrate::mail::{Source, SourceAddr};
-
-        let registry = Arc::new(Registry::new());
-        for d in descriptors::all() {
-            let _ = registry.register_kind_with_descriptor(d);
-        }
-        let mailer = Arc::new(Mailer::new(Arc::clone(&registry)));
-
-        // Wire one settlement registry into both seams (the chassis builder
-        // does both installs at boot) so an armed Call drains cleanly.
-        let settlement = Arc::new(SettlementRegistry::new());
-        mailer.install_settlement_registry(Arc::clone(&settlement));
-        mailer
-            .trace_handle()
-            .install_settlement_registry(Arc::clone(&settlement));
-
-        // A `Component` caller inbox captures the reply so the test reads
-        // its delivered `root`.
-        let (reply_tx, reply_rx) = mpsc::channel::<OwnedDispatch>();
-        let caller_mailbox = registry.register_inbox(
-            "test.window.log_tail.caller",
-            Arc::new(move |dispatch: OwnedDispatch| {
-                dispatch.discharge();
-                let _ = reply_tx.send(dispatch);
-            }) as Arc<dyn InboxHandler>,
-        );
-
-        // The window inbox forwards armed envelopes onto the
-        // `SettlingInbox`'s channel, exactly as `claim_mailbox` does.
-        let window_mailbox = mailbox_id_from_name(
-            <aether_capabilities::HeadlessWindowCapability as Addressable>::NAMESPACE,
-        );
-        let (tx, rx) = mpsc::channel::<Envelope>();
-        let handler: Arc<dyn InboxHandler> = Arc::new(move |d: Envelope| {
-            let _ = tx.send(d);
-        });
-        registry
-            .try_register_inbox_with_id(window_mailbox, "aether.window", handler)
-            .expect("register the window inbox");
-        let inbox = SettlingInbox::new(window_mailbox, rx, Arc::clone(&mailer));
-
-        // Push a real armed `LogTail` Call whose reply target is the
-        // caller inbox, then drain it to an `InboundMail` guard.
-        let root = MailId::new(window_mailbox, 1);
-        let mail_id = MailId::new(window_mailbox, 2);
-        mailer.record_sent_inflight(root);
-        let caller_source = Source::with_correlation(SourceAddr::Component(caller_mailbox), 0x99);
-        let bytes = LogTail {
-            max: 8,
-            min_level: None,
-            since: None,
-        }
-        .encode_into_bytes();
-        mailer.push(
-            Mail::new(window_mailbox, <LogTail as Kind>::ID, bytes, 1)
-                .with_reply_to(caller_source)
-                .with_lineage(mail_id, root, None),
-        );
-        let mail = inbox.try_next().expect("the armed LogTail Call is queued");
-
-        let slots = ActorSlots::new();
-        let matched = with_stamped(&slots, || {
-            try_framework_dispatch(&mailer, window_mailbox, &mail)
-        });
-        assert!(
-            matched,
-            "framework dispatch arm must match a LogTail Call at aether.window",
-        );
-
-        let dispatch = reply_rx
-            .recv_timeout(Duration::from_secs(2))
-            .expect("framework arm routed a reply to the caller inbox");
-        assert_eq!(
-            dispatch.kind_name,
-            <LogTailResult as Kind>::NAME,
-            "the reply is a LogTailResult",
-        );
-        assert_eq!(
-            dispatch.root, root,
-            "the framework-arm reply joins the inbound's causal chain (#1710)",
-        );
-
-        // Drop the guard last so its `Finished` records after the reply's
-        // `Sent` (ADR-0080 §6) — settlement bookkeeping stays balanced.
-        drop(mail);
-    }
-
-    /// A non-framework kind (here `SetWindowTitle`) does NOT trip the
-    /// framework arms — the driver-specific path keeps its turn so
-    /// `actor_logs`-style queries don't shadow the existing window
-    /// controls. The helpers return `None` and the driver dispatch
-    /// continues; nothing is replied here.
-    #[test]
-    fn try_framework_dispatch_skips_window_kinds() {
-        use std::sync::mpsc;
-
-        use aether_actor::local::ActorSlots;
-        use aether_data::MailId;
-        use aether_kinds::descriptors;
-        use aether_substrate::actor::native::local::with_stamped;
-
-        use aether_substrate::mail::Mail;
-        use aether_substrate::mail::registry::{InboxHandler, Registry};
-
-        let registry = Arc::new(Registry::new());
-        for d in descriptors::all() {
-            let _ = registry.register_kind_with_descriptor(d);
-        }
-        let mailer = Arc::new(Mailer::new(Arc::clone(&registry)));
-
-        let window_mailbox = mailbox_id_from_name(
-            <aether_capabilities::HeadlessWindowCapability as Addressable>::NAMESPACE,
-        );
-        let (tx, rx) = mpsc::channel::<Envelope>();
-        let handler: Arc<dyn InboxHandler> = Arc::new(move |d: Envelope| {
-            let _ = tx.send(d);
-        });
-        registry
-            .try_register_inbox_with_id(window_mailbox, "aether.window", handler)
-            .expect("register the window inbox");
-        let inbox = SettlingInbox::new(window_mailbox, rx, Arc::clone(&mailer));
-
-        // A `MailId::NONE` push keeps the drained guard disarmed (no armed
-        // Call to settle) — the test pins only the skip verdict.
-        let payload = SetWindowTitle {
-            title: "ignored".to_owned(),
-        }
-        .encode_into_bytes();
-        mailer.push(
-            Mail::new(window_mailbox, <SetWindowTitle as Kind>::ID, payload, 1).with_lineage(
-                MailId::NONE,
-                MailId::NONE,
-                None,
-            ),
-        );
-        let mail = inbox.try_next().expect("the SetWindowTitle push is queued");
-
-        let slots = ActorSlots::new();
-        let matched = with_stamped(&slots, || {
-            try_framework_dispatch(&mailer, window_mailbox, &mail)
-        });
-        assert!(!matched, "SetWindowTitle is a driver-specific kind");
-    }
-
-    /// iamacoffeepot/aether#1325, re-homed on ADR-0106: the window inbox
-    /// drain owns the ADR-0080 §2 settlement bracket for every inbound
-    /// envelope (the `Inbox` mailbox records none on the producer side),
-    /// now by construction — draining a real armed Call through a
-    /// `SettlingInbox` and letting the `InboundMail` guard fall out of
-    /// scope settles the root and disarms the ADR-0094 guard (no #1704
-    /// abort). The CI-runnable regression guard for the window drain
-    /// without standing up winit/wgpu; the windowed end-to-end
-    /// blocking-send path stays MCP-manual.
-    #[test]
-    fn window_inbox_drain_settles_root_on_guard_drop() {
-        use std::sync::mpsc;
-
-        use aether_data::{Kind, MailId};
-        use aether_kinds::descriptors;
-
-        use aether_substrate::chassis::settlement::SettlementRegistry;
-        use aether_substrate::mail::Mail;
-        use aether_substrate::mail::registry::{InboxHandler, Registry};
-
-        fn title_payload() -> Vec<u8> {
-            SetWindowTitle {
-                title: "ignored".to_owned(),
-            }
-            .encode_into_bytes()
-        }
-
-        let registry = Arc::new(Registry::new());
-        for d in descriptors::all() {
-            let _ = registry.register_kind_with_descriptor(d);
-        }
-        let mailer = Arc::new(Mailer::new(Arc::clone(&registry)));
-
-        // Wire a settlement registry into both seams (the chassis builder
-        // does both installs at boot, builder.rs:1119-1122) so the
-        // emit-time counter's zero-transition can `fire_settled`.
-        let settlement = Arc::new(SettlementRegistry::new());
-        mailer.install_settlement_registry(Arc::clone(&settlement));
-        mailer
-            .trace_handle()
-            .install_settlement_registry(Arc::clone(&settlement));
-
-        // Register the window mailbox forwarding armed envelopes onto the
-        // `SettlingInbox`'s channel, exactly as `claim_mailbox` does.
-        let window_mailbox = mailbox_id_from_name(
-            <aether_capabilities::HeadlessWindowCapability as Addressable>::NAMESPACE,
-        );
-        let (tx, rx) = mpsc::channel::<Envelope>();
-        let handler: Arc<dyn InboxHandler> = Arc::new(move |d: Envelope| {
-            let _ = tx.send(d);
-        });
-        registry
-            .try_register_inbox_with_id(window_mailbox, "aether.window", handler)
-            .expect("register the window inbox");
-        let inbox = SettlingInbox::new(window_mailbox, rx, Arc::clone(&mailer));
-
-        // A real armed Call: seed the root, push through the mailer so the
-        // `route_mail` Inbox arm arms the obligation guard with `mail_id`,
-        // drain via `try_next`, then drop — the guard's `Drop` records
-        // `Finished` and disarms, settling the root.
-        let root = MailId::new(window_mailbox, 1);
-        let mail_id = MailId::new(window_mailbox, 2);
-        mailer.record_sent_inflight(root);
-        let settle = settlement.subscribe_settlement(root);
-        mailer.push(
-            Mail::new(
-                window_mailbox,
-                <SetWindowTitle as Kind>::ID,
-                title_payload(),
-                1,
-            )
-            .with_lineage(mail_id, root, None),
-        );
-        drop(inbox.try_next().expect("the armed Call is queued"));
-        settle
-            .recv()
-            .expect("window root settles when the drained mail's guard drops");
-
-        // A `MailId::NONE` push (window-size / frame-stats) settles
-        // nothing — the guard mints disarmed and `record_finished` no-ops.
-        let guard_root = MailId::new(window_mailbox, 3);
-        mailer.record_sent_inflight(guard_root);
-        let guard_rx = settlement.subscribe_settlement(guard_root);
-        mailer.push(
-            Mail::new(
-                window_mailbox,
-                <SetWindowTitle as Kind>::ID,
-                title_payload(),
-                1,
-            )
-            .with_lineage(MailId::NONE, guard_root, None),
-        );
-        drop(inbox.try_next().expect("the NONE push is queued"));
-        assert!(
-            guard_rx.try_recv().is_err(),
-            "a NONE inbound discharges no root",
-        );
-    }
-
-    /// iamacoffeepot/aether#1704: the lifecycle reply inbox is a
-    /// hand-rolled `claim_mailbox` consumer, so it must run the ADR-0094
-    /// obligation + ADR-0080 §2 settlement bracket itself — the sibling
-    /// window arm's #1325 fix, applied at the reply consume site. Route a
-    /// real, obligation-**armed** `LifecycleAdvanceComplete` reply through
-    /// the registered inbox (the production `route_mail` Inbox arm arms the
-    /// guard), then drive `consume_lifecycle_reply` over it and assert the
-    /// guard is disarmed (no abort on drop) AND the settlement counter
-    /// balances — on both a non-`NONE`-root reply (the discharge is the
-    /// sole `Finished`, so the root settles) and the `NONE`-root per-frame
-    /// reply (a counter no-op). Pre-#1704 the armed guard aborted the
-    /// process when the consume site dropped the envelope without
-    /// `discharge`.
-    #[test]
-    fn consume_lifecycle_reply_discharges_armed_reply_and_balances_settlement() {
-        use std::sync::mpsc;
-
-        use aether_data::MailId;
-
-        use aether_substrate::chassis::settlement::SettlementRegistry;
-        use aether_substrate::mail::registry::{InboxHandler, Registry};
-
-        let registry = Arc::new(Registry::new());
-        let mailer = Arc::new(Mailer::new(Arc::clone(&registry)));
-
-        // Wire one settlement registry into both seams (the chassis builder
-        // does both installs at boot) so the counter's zero-transition can
-        // `fire_settled`.
-        let settlement = Arc::new(SettlementRegistry::new());
-        mailer.install_settlement_registry(Arc::clone(&settlement));
-        mailer
-            .trace_handle()
-            .install_settlement_registry(Arc::clone(&settlement));
-
-        // Register the reply inbox exactly as `claim_mailbox` does: forward
-        // the obligation-armed envelope onto the `SettlingInbox`'s channel,
-        // carrying its guard with it so the framework drain owns the
-        // discharge.
-        let (tx, rx) = mpsc::channel::<Envelope>();
-        let handler: Arc<dyn InboxHandler> = Arc::new(move |dispatch: Envelope| {
-            let _ = tx.send(dispatch);
-        });
-        let reply_mailbox = registry
-            .try_register_inbox("aether.lifecycle.advance_reply", handler)
-            .expect("register the reply inbox");
-        let inbox = SettlingInbox::new(reply_mailbox, rx, Arc::clone(&mailer));
-
-        let cap_mailbox = mailbox_id_from_name(
-            <aether_capabilities::LifecycleCapability as Addressable>::NAMESPACE,
-        );
-
-        // (1) Non-`NONE`-root reply (the degraded `on_advance` inline-reply
-        // shape): the producer hook records the reply's `Sent` against
-        // `root`, and the `InboundMail` guard's `Drop` inside
-        // `consume_lifecycle_reply` is the sole `Finished`, so the root
-        // settles.
-        let root = MailId::new(cap_mailbox, 1);
-        // 1<<63 is the disjoint reply-lineage base (#1701) — a non-`NONE`
-        // mail_id, so the real Inbox arm arms the obligation guard.
-        let armed_reply_id = MailId::new(cap_mailbox, 1 << 63);
-        let settle_rx = settlement.subscribe_settlement(root);
-        let sender = Source::with_correlation(SourceAddr::Component(reply_mailbox), 7);
-        mailer.send_reply(
-            sender,
-            &LifecycleAdvanceComplete {
-                completed: 1,
-                next: 42,
-            },
-            armed_reply_id,
-            root,
-            None,
-        );
-        let mail = inbox.try_next().expect("armed reply routed to the inbox");
-        match consume_lifecycle_reply(mail) {
-            LifecycleReplyOutcome::Complete(next) => {
-                assert_eq!(next, Some(42), "decodes the advance-complete `next`");
-            }
-            LifecycleReplyOutcome::Unexpected => panic!("expected the advance-complete arm"),
-        }
-        settle_rx
-            .recv()
-            .expect("the guard's Finished balances the reply's Sent and settles the root");
-
-        // (2) `NONE`-root reply (the real per-frame deferred path, replying
-        // to a bare lineage-less `Settled` notice): the producer's
-        // `record_sent_inflight` no-ops, the drop's `record_finished` is a
-        // counter no-op, and the armed guard is still disarmed so the
-        // envelope drops without the ADR-0094 abort. Reaching the assert at
-        // all proves the guard was disarmed.
-        let armed_reply_id_2 = MailId::new(cap_mailbox, (1 << 63) + 1);
-        let sender_2 = Source::with_correlation(SourceAddr::Component(reply_mailbox), 8);
-        mailer.send_reply(
-            sender_2,
-            &LifecycleAdvanceComplete {
-                completed: 2,
-                next: 0,
-            },
-            armed_reply_id_2,
-            MailId::NONE,
-            None,
-        );
-        let mail_2 = inbox
-            .try_next()
-            .expect("second armed reply routed to the inbox");
-        match consume_lifecycle_reply(mail_2) {
-            LifecycleReplyOutcome::Complete(next) => {
-                assert_eq!(next, Some(0), "the terminal reply decodes `next == 0`");
-            }
-            LifecycleReplyOutcome::Unexpected => panic!("expected the advance-complete arm"),
-        }
-    }
-
-    /// iamacoffeepot/aether#1317: the occluded-capture branch selection.
-    /// This is the CI-runnable core of the fail-fast — the winit wiring
-    /// (`fail_capture_if_occluded` → `take` → `reply.reply` → drop) stays
-    /// MCP-manual because `ControlFlow::Wait` + `RedrawRequested`
-    /// suppression has no CI display surface. Here we pin only the pure
-    /// branch logic: occluded-with-parked-capture selects an `Err` reply,
-    /// occluded-with-empty-slot and visible-window are no-ops/redraw.
-    #[test]
-    fn occluded_capture_disposition_selects_failfast_only_when_occluded_and_parked() {
-        use std::sync::mpsc;
-
-        use aether_data::{KindId, MailId};
-        use aether_kinds::trace::Nanos;
-        use aether_substrate::capture::PendingCapture;
-        use aether_substrate::mail::MailRef;
-        use aether_substrate::mail::Source;
-        use aether_substrate::mail::registry::{OwnedDispatch, Registry};
-
-        fn parked() -> PendingCapture {
-            // A NONE-lineage disarmed inbound: its retained guard carries no
-            // settlement obligation, so dropping it is a clean no-op and
-            // this branch-selection test needs no settlement registry. Sent
-            // straight onto the `SettlingInbox`'s channel, then drained to
-            // the guard.
-            let registry = Arc::new(Registry::new());
-            let mailer = Arc::new(Mailer::new(registry));
-            let mailbox = MailboxId(0x0CAB);
-            let (tx, rx) = mpsc::channel::<Envelope>();
-            tx.send(OwnedDispatch::disarmed(
-                KindId(0),
-                "test.capture.parked".to_owned(),
-                None,
-                Source::NONE,
-                MailRef::from(Vec::new()),
-                1,
-                MailId::NONE,
-                MailId::NONE,
-                None,
-                Nanos(0),
-                0,
-                mailbox,
-            ))
-            .expect("queue the parked inbound");
-            let inbox = SettlingInbox::new(mailbox, rx, mailer);
-            let reply = inbox.try_next().expect("one queued");
-            PendingCapture {
-                reply,
-                after_mails: Vec::new(),
-                pre_settlements: Vec::new(),
-                checks: Vec::new(),
-                reference: None,
-            }
-        }
-
-        // Occluded + a parked capture → fail it fast with an `Err` reply
-        // whose message names `aether.window.focus` as the remedy.
-        match occluded_capture_disposition(true, Some(parked())) {
-            OccludedCaptureDisposition::FailFast { result, .. } => match result {
-                CaptureFrameResult::Err { error } => {
-                    assert!(
-                        error.contains("occluded") && error.contains("aether.window.focus"),
-                        "Err reply must name occlusion + the focus remedy, got: {error}",
-                    );
-                }
-                CaptureFrameResult::Ok { .. } => panic!("occluded capture must fail, not Ok"),
-            },
-            _ => panic!("occluded + parked capture must select FailFast"),
-        }
-
-        // Occluded but nothing parked → no-op (already serviced / stale wake).
-        assert!(
-            matches!(
-                occluded_capture_disposition(true, None),
-                OccludedCaptureDisposition::Empty
-            ),
-            "occluded + empty slot must be a no-op",
-        );
-
-        // Visible window → fall through to redraw, regardless of the slot.
-        assert!(
-            matches!(
-                occluded_capture_disposition(false, Some(parked())),
-                OccludedCaptureDisposition::Redraw
-            ),
-            "visible window must fall through to redraw",
-        );
-        assert!(
-            matches!(
-                occluded_capture_disposition(false, None),
-                OccludedCaptureDisposition::Redraw
-            ),
-            "visible window must fall through to redraw",
-        );
-    }
-
-    /// iamacoffeepot/aether#1758: the deferred capture reply joins the
-    /// inbound's ADR-0080 causal chain through the `InboundMail` guard the
-    /// render cap parked on `PendingCapture` (via `take_inbound`). Replying
-    /// through `req.reply.reply` records the reply's `Sent`; dropping `req`
-    /// records the inbound's `Finished` *after* it (ADR-0080 §6), so the
-    /// caller's root stays open until the reply itself finishes — settling
-    /// exactly once. This pins the capture-migration's reply-before-drop
-    /// order without standing up wgpu/winit, mirroring the claimed-inbox
-    /// `reply_sent_recorded_before_inbound_finished` shape through the
-    /// capture queue's parked request.
-    #[test]
-    fn capture_reply_joins_held_chain() {
-        use std::sync::mpsc;
-
-        use aether_data::{Kind, MailId};
-        use aether_kinds::{CaptureFrame, CaptureFrameResult, descriptors};
-
-        use aether_substrate::capture::PendingCapture;
-        use aether_substrate::chassis::settlement::SettlementRegistry;
-        use aether_substrate::mail::registry::{InboxHandler, Registry};
-        use aether_substrate::mail::{Mail, Source, SourceAddr};
-
-        let registry = Arc::new(Registry::new());
-        for d in descriptors::all() {
-            let _ = registry.register_kind_with_descriptor(d);
-        }
-        let mailer = Arc::new(Mailer::new(Arc::clone(&registry)));
-
-        // One settlement registry wired into both seams (the chassis builder
-        // does both installs at boot) so the inbound + reply settle cleanly.
-        let settlement = Arc::new(SettlementRegistry::new());
-        mailer.install_settlement_registry(Arc::clone(&settlement));
-        mailer
-            .trace_handle()
-            .install_settlement_registry(Arc::clone(&settlement));
-
-        // A `Component` reply target captures the deferred reply so the test
-        // reads its delivered `root` and finishes it — the desktop MCP
-        // capture reply target is a `Component(rpc_server)`.
-        let (reply_tx, reply_rx) = mpsc::channel::<Envelope>();
-        let target = registry.register_inbox(
-            "test.capture.reply_target",
-            Arc::new(move |d: Envelope| {
-                let _ = reply_tx.send(d);
-            }) as Arc<dyn InboxHandler>,
-        );
-
-        // Push a real armed `CaptureFrame` Call whose reply target is the
-        // captured inbox, then drain it to the `PendingCapture`'s retained
-        // guard (the render cap's `take_inbound`, mirrored out-of-crate).
-        let render_mailbox = MailboxId(0x0CA6);
-        let (tx, rx) = mpsc::channel::<Envelope>();
-        let handler: Arc<dyn InboxHandler> = Arc::new(move |d: Envelope| {
-            let _ = tx.send(d);
-        });
-        registry
-            .try_register_inbox_with_id(render_mailbox, "test.capture.render", handler)
-            .expect("register the render inbox");
-        let inbox = SettlingInbox::new(render_mailbox, rx, Arc::clone(&mailer));
-
-        let root = MailId::new(render_mailbox, 1);
-        let mail_id = MailId::new(render_mailbox, 2);
-        mailer.record_sent_inflight(root);
-        let settle = settlement.subscribe_settlement(root);
-        let caller = Source::with_correlation(SourceAddr::Component(target), 0x42);
-        mailer.push(
-            Mail::new(render_mailbox, <CaptureFrame as Kind>::ID, Vec::new(), 1)
-                .with_reply_to(caller)
-                .with_lineage(mail_id, root, None),
-        );
-        let reply = inbox
-            .try_next()
-            .expect("the armed CaptureFrame Call is queued");
-        let req = PendingCapture {
-            reply,
-            after_mails: Vec::new(),
-            pre_settlements: Vec::new(),
-            checks: Vec::new(),
-            reference: None,
-        };
-
-        // The driver's reply path: reply through the retained guard, then
-        // let `req` (the parked request) drop.
-        assert!(
-            req.reply.reply(&CaptureFrameResult::Err {
-                error: "deferred".to_owned(),
-            }),
-            "the deferred capture reply routed to the Component target",
-        );
-        assert!(
-            settle.try_recv().is_err(),
-            "the reply's Sent holds the chain open",
-        );
-        drop(req);
-        assert!(
-            settle.try_recv().is_err(),
-            "inbound Finished alone does not settle — the deferred reply is still in flight",
-        );
-
-        // Finish the reply the way the RPC server's dispatcher would; only
-        // now does the root settle — exactly once.
-        let reply_env = reply_rx
-            .recv()
-            .expect("deferred reply routed to the target");
-        assert_eq!(
-            reply_env.kind_name,
-            <CaptureFrameResult as Kind>::NAME,
-            "the deferred reply is a CaptureFrameResult",
-        );
-        assert_eq!(
-            reply_env.root, root,
-            "the deferred capture reply joins the inbound's causal chain (#1758)",
-        );
-        let reply_id = reply_env.mail_id;
-        reply_env.discharge();
-        mailer.record_finished(reply_id, root);
-        settle
-            .recv()
-            .expect("root settles once the deferred capture reply finishes");
     }
 }
